Enable tap to turn page back and forth via genie
