+++ argreader.py
@@ -1,0 +1,4 @@
+import sys
+import swipe
+
+swipe.swipe(sys.argv[1])
