+++ genie.example.conf
@@ -1,0 +1,9 @@
+gesture=tap
+command=/home/root/pdfswiper/runner.sh right
+zone=0.15 0.8 0.4 1
+fingers=1
+
+gesture=tap
+command=/home/root/pdfswiper/runner.sh left
+zone=0 0.8 0.15 1
+fingers=1
